implement f64 version of sphere intersection

+++ rust/bevy_experiment/src/orbit_camera/config.rs
@@ -68,7 +68,7 @@ pub struct OrbitCameraConfig {
     // pub scroll_line_sensitivity: f32,
     // /// For devices with smooth scrolling, like touchpads
     // pub scroll_pixel_sensitivity: f32,
-    pub earth_radius: f32,
+    pub earth_radius: f64,
 
     /// Latitude threshold below which roll is fully constrained (degrees)
     pub roll_constraint_low_lat: f64,
